Make stepping work for rounded colliders
Stepping up and down now works for capsules and spheres.
The radius of capsule and sphere colliders are now taken into consideration when checking if a floor surface is not too high.
Snapping condition now checks if a floor collision occoured during sliding, instead of checking ```is_on_floor```, which can be set to true when checking pre move collisions.. This makes snapping down stairs work for rounded shapes.