Modify stationary collision check and snap logic
Fixes #6. A condition is added to the stationary/motionless collision check that disables the floor status update, when previously had a floor collider saved, or when detaching from a platform.
Snapping now works on upwards moving platforms, floor data is now acquired with snapping on upwards woving platforms, not with the stationary check.